presupuesto-backend/app/core/security.py
from datetime import datetime, timedelta
from typing import Dict, Optional

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from app.models.auth_dto import UserAD


from app.core.config_core import settings

# Configuración de esquema de seguridad
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/login")


def create_access_token(data: Dict, expires_delta: Optional[int] = None) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=expires_delta or settings.JWT_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.JWT_SECRET, algorithm="HS256")
    return encoded_jwt


def get_current_user(token: str = Depends(oauth2_scheme)) -> UserAD:
    try:
        payload = jwt.decode(token, settings.JWT_SECRET, algorithms=["HS256"])
        user = UserAD(
            sAMAccountName=payload.get("sub"),
            displayName=payload.get("displayName"),
            groups=payload.get("groups", [])
        )
        return user
    except JWTError as e:
        raise HTTPException(status_code=401, detail="Token inválido")
